+++ ddStringTools_snippet.php
@@ -1,169 +1,21 @@
 <?php
 /**
  * ddStringTools
- * @version 1.7 (2020-06-22)
+ * @version 2.0 (2021-04-15)
  * 
  * @see README.md
  * 
- * @copyright 2016–2020 DD Group {@link http://DivanDesign.biz }
+ * @copyright 2016–2021 DD Group {@link https://DivanDesign.biz }
  */
-
-$snippetPath =
-	$modx->getConfig('base_path') .
-	implode(
-		DIRECTORY_SEPARATOR,
-		[
-			'assets',
-			'snippets',
-			'ddStringTools',
-			''
-		]
-	)
-;
-
-$snippetPath_src_tool =
-	$snippetPath .
-	implode(
-		DIRECTORY_SEPARATOR,
-		[
-			'src',
-			'Tool',
-			''
-		]
-	)
-;
 
 //Include (MODX)EvolutionCMS.libraries.ddTools
 require_once(
 	$modx->getConfig('base_path') .
-	implode(
-		DIRECTORY_SEPARATOR,
-		[
-			'assets',
-			'libs',
-			'ddTools',
-			'modx.ddtools.class.php'
-		]
-	)
+	'assets/libs/ddTools/modx.ddtools.class.php'
 );
 
-if(!class_exists('\ddStringTools\Tool\Tool')){
-	require_once(
-		$snippetPath .
-		'require.php'
-	);
-}
-
-if (!isset($inputString)){
-	$inputString = '';
-}else{
-	//If the input string is passed as an object (e. g. through `$modx->runSnippet`)
-	if (
-		is_object($inputString) ||
-		is_array($inputString)
-	){
-		//Convert it to JSON
-		$inputString = \DDTools\ObjectTools::convertType([
-			'object' => $inputString,
-			'type' => 'stringJsonAuto'
-		]);
-	}
-}
-
-//Backward compatibility
-if (!isset($tools)){
-	//Simple boolean params
-	$tools = \ddTools::verifyRenamedParams([
-		'params' => $params,
-		'compliance' => [
-			'specialCharConverter' => 'specialCharsToHTMLEntities',
-			'placeholderRemover' => 'removePlaceholders',
-			'charEscaper' => 'escapeForJS',
-			'urlEncoder' => 'URLEncode'
-		],
-		'writeToLog' => false
-	]);
-	
-	//caseConverter
-	if (isset($toUppercase)){
-		$tools['caseConverter'] = [
-			'toUpper' => boolval($toUppercase)
-		];
-	}else if (isset($toLowercase)){
-		$tools['caseConverter'] = [
-			'toLower' => boolval($toLowercase)
-		];
-	}
-	
-	//markdownParser
-	if (isset($parseMarkdown)){
-		if ($parseMarkdown == 'line'){
-			$tools['markdownParser'] = [
-				'parseInline' => true
-			];
-		}else{
-			$tools['markdownParser'] = true;
-		}
-	}
-	
-	//typographer
-	if (
-		isset($typography) &&
-		$typography == '1'
-	){
-		if (isset($typography_params)){
-			$tools['typographer'] = \DDTools\ObjectTools::convertType([
-				'object' => $typography_params,
-				'type' => 'objectStdClass'
-			]);
-		}else{
-			$tools['typographer'] = true;
-		}
-	}
-	
-	//tagRemover
-	if (
-		isset($stripTags) &&
-		$stripTags == '1'
-	){
-		if (isset($stripTags_allowed)){
-			$tools['tagRemover'] = [
-				'allowed' => $stripTags_allowed
-			];
-		}else{
-			$tools['tagRemover'] = true;
-		}
-	}
-	
-	if (!empty($tools)){
-		\ddTools::logEvent([
-			'message' => '<p>The snippet parameters were changed, please see the documentation and correct your calls.</p>'
-		]);
-	}
-}else{
-	$tools = \DDTools\ObjectTools::convertType([
-		'object' => $tools,
-		'type' => 'objectStdClass'
-	]);
-}
-
-foreach (
-	$tools as
-	$toolName =>
-	$toolParams
-){
-	//Senselessly to process empty strings. We need to check this on each cycle iteration because string can become empty after one of iterations.
-	if ($inputString != ''){
-		$toolObject = \ddStringTools\Tool\Tool::createChildInstance([
-			'name' => $toolName,
-			'parentDir' => $snippetPath_src_tool,
-			//Passing parameters into constructor
-			'params' => $toolParams
-		]);
-		
-		$inputString = $toolObject->modify($inputString);
-	}
-}
-
-return $inputString;
+return \DDTools\Snippet::runSnippet([
+	'name' => 'ddStringTools',
+	'params' => $params
+]);
 ?>