+++ require.php
@@ -1,12 +1,14 @@
 <?php
+//Include (MODX)EvolutionCMS.libraries.ddTools
 require_once(
-	implode(
-		DIRECTORY_SEPARATOR,
-		[
-			'src',
-			'Tool',
-			'Tool.php',
-		]
-	)
+	//path to `assets`
+	dirname(
+		__DIR__,
+		2
+	) .
+	'/libs/ddTools/modx.ddtools.class.php'
 );
+
+require_once('src/Tool/Tool.php');
+require_once('src/Snippet.php');
 ?>